auto works, normal position for shoulder, deleted auto specimen+sample
