+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modules/Shoulder.java
@@ -17,9 +17,9 @@ public class Shoulder {
     public static double SH_POS_MIN = 0;
     public static double SHOULDER_STEP = 0.005;
 
-    public static double SH_POS_TO_BASKET = .81;// 0.5561; upper
+    public static double SH_POS_TO_BASKET = .89;// 0.5561; upper
 
-    public static double SH_POS_TO_BASKET_AUTO = .8;
+    public static double SH_POS_TO_BASKET_AUTO = .86;
     public static double SH_POS_TO_INTAKE = 0;
 
     public static double SH_POS_INIT = 0.2;
