+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/disabled/Auto3Basket.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes.auto;
+package org.firstinspires.ftc.teamcode.opmodes.auto.disabled;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
@@ -9,6 +9,7 @@ import org.firstinspires.ftc.teamcode.modules.driveTrainMecanum.DriveTrainMecanu
 import org.firstinspires.ftc.teamcode.roadrunner.trajectorysequence.TrajectorySequence;
 
 @Autonomous(name = "2 + 1 Auto Basket", group = "Robot")
+@Disabled
 public class Auto3Basket extends LinearOpMode {
 
     @Override
